// public/frontend/src/app/utensils/sort.enum.ts
export enum SortOrder {
  None = 0,
  PriceAsc = 'priceAsc',
  PriceDesc = 'priceDesc',
  ColourAsc = 'colourAsc',
  ColourDesc = 'colourDesc',
  MaterialAsc = 'materialAsc',
  MaterialDesc = 'materialDesc',
  TypeAsc = 'typeAsc',
  TypeDesc = 'typeDesc',
  SizeAsc = 'sizeAsc',
  SizeDesc = 'sizeDesc',
  ConditionAsc = 'conditionAsc',
  ConditionDesc = 'conditionDesc'
}